rename obfuscation package to signals, update variable names accordingly
