+++ internal/staticanalysis/parsing/analyze.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"github.com/ossf/package-analysis/internal/staticanalysis/externalcmd"
-	"github.com/ossf/package-analysis/internal/staticanalysis/obfuscation/stringentropy"
+	"github.com/ossf/package-analysis/internal/staticanalysis/signals/stringentropy"
 	"github.com/ossf/package-analysis/internal/staticanalysis/token"
 )
