+++ internal/staticanalysis/signals/detections/escape_sequences.go
