+++ internal/staticanalysis/signals/detections/escape_sequences_test.go
