+++ internal/staticanalysis/signals/detections/hex_strings.go
