+++ internal/staticanalysis/signals/detections/hex_strings_test.go
