+++ internal/staticanalysis/signals/detections/suspicious_identifiers.go
