+++ internal/staticanalysis/signals/stringentropy/string_entropy_test.go
